+++ lib/measures/openstudio_results/measure.rb
@@ -36,9 +36,11 @@
 require 'erb'
 require 'json'
 
+# load local resources
 require "#{File.dirname(__FILE__)}/resources/os_lib_reporting"
 require "#{File.dirname(__FILE__)}/resources/os_lib_schedules"
 require "#{File.dirname(__FILE__)}/resources/os_lib_helper_methods"
+
 require_relative 'resources/Siz.Model'
 
 # start the measure
@@ -278,45 +280,43 @@ class OpenStudioResults < OpenStudio::Measure::ReportingMeasure
       # generate data for requested sections
       sections_made = 0
       possible_sections.each do |method_name|
-        begin
-          next unless args[method_name]
-          section = false
-          eval("section = OsLib_Reporting.#{method_name}(model,sql_file,runner,false,is_ip_units)")
-          display_name = eval("OsLib_Reporting.#{method_name}(nil,nil,nil,true)[:title]")
-          if section
-            @sections << section
-            sections_made += 1
-            # look for empty tables and warn if skipped because returned empty
-            section[:tables].each do |table|
-              if !table
-                runner.registerWarning("A table in #{display_name} section returned false and was skipped.")
-                section[:messages] = ["One or more tables in #{display_name} section returned false and was skipped."]
-              end
+        next unless args[method_name]
+        section = false
+        eval("section = OsLib_Reporting.#{method_name}(model,sql_file,runner,false,is_ip_units)")
+        display_name = eval("OsLib_Reporting.#{method_name}(nil,nil,nil,true)[:title]")
+        if section
+          @sections << section
+          sections_made += 1
+          # look for empty tables and warn if skipped because returned empty
+          section[:tables].each do |table|
+            if !table
+              runner.registerWarning("A table in #{display_name} section returned false and was skipped.")
+              section[:messages] = ["One or more tables in #{display_name} section returned false and was skipped."]
             end
-          else
-            runner.registerWarning("#{display_name} section returned false and was skipped.")
-            section = {}
-            section[:title] = display_name.to_s
-            section[:tables] = []
-            section[:messages] = []
-            section[:messages] << "#{display_name} section returned false and was skipped."
-            @sections << section
           end
-        rescue StandardError => e
-          display_name = eval("OsLib_Reporting.#{method_name}(nil,nil,nil,true)[:title]")
-          if display_name.nil? then display_name == method_name end
-          runner.registerWarning("#{display_name} section failed and was skipped because: #{e}. Detail on error follows.")
-          runner.registerWarning(e.backtrace.join("\n").to_s)
-
-          # add in section heading with message if section fails
+        else
+          runner.registerWarning("#{display_name} section returned false and was skipped.")
           section = {}
           section[:title] = display_name.to_s
           section[:tables] = []
           section[:messages] = []
-          section[:messages] << "#{display_name} section failed and was skipped because: #{e}. Detail on error follows."
-          section[:messages] << [e.backtrace.join("\n").to_s]
+          section[:messages] << "#{display_name} section returned false and was skipped."
           @sections << section
         end
+      rescue StandardError => e
+        display_name = eval("OsLib_Reporting.#{method_name}(nil,nil,nil,true)[:title]")
+        if display_name.nil? then display_name == method_name end
+        runner.registerWarning("#{display_name} section failed and was skipped because: #{e}. Detail on error follows.")
+        runner.registerWarning(e.backtrace.join("\n").to_s)
+
+        # add in section heading with message if section fails
+        section = {}
+        section[:title] = display_name.to_s
+        section[:tables] = []
+        section[:messages] = []
+        section[:messages] << "#{display_name} section failed and was skipped because: #{e}. Detail on error follows."
+        section[:messages] << [e.backtrace.join("\n").to_s]
+        @sections << section
       end
 
     else