+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # OpenStudio Common Measures Gem
 
-## Version 0.X (unreleased)
+## Version 0.2.0 (Unreleased)
 
+* Upgrade Bundler to 2.1.x
+* Restrict to Ruby ~> 2.5.0
+* Remove json_pure gem
+* Update openstudio-extension to 0.2.2
+* Update measure tester to 0.2.0 (removes need for github checkout)
 * Exclude measure tests from being released with the gem (reduces the size of the installed gem significantly) 
 
 ## Version 0.1.2